move indexer to a proper module
