+++ backend/setup.py
@@ -92,7 +92,7 @@ setup(
             "inspirehep_migrator = inspirehep.migrator.models",
         ],
         "invenio_celery.tasks": [
-            "inspirehep_indexer = inspirehep.records.indexer.tasks",
+            "inspirehep_indexer = inspirehep.indexer.tasks",
             "inspirehep_migrator = inspirehep.migrator.tasks",
             "inspirehep_orcid = inspirehep.orcid.tasks",
             "inspirehep_submissions = inspirehep.submissions.tasks",
